Refactor vault credential secret handling
