+++ internal/vaults/handler.go
@@ -8,8 +8,6 @@ import (
 	"fmt"
 	"log/slog"
 	"net/http"
-	"net/url"
-	"regexp"
 	"strconv"
 	"strings"
 	"time"
@@ -28,8 +26,6 @@ import (
 )
 
 const maxVaultBodySize = 4 << 20
-
-var credentialHostPattern = regexp.MustCompile(`^(\*\.)?[A-Za-z0-9.-]+$`)
 
 type Handler struct {
 	cfg       config.Config
@@ -109,89 +105,6 @@ type validationHTTPResponse struct {
 	StatusCode    int    `json:"status_code"`
 }
 
-type credentialAuthState struct {
-	AuthType      string
-	Key           string
-	PublicAuth    json.RawMessage
-	SecretPayload json.RawMessage
-}
-
-var (
-	// ErrMissingSecretEnvelope is returned when an active credential has no
-	// secret envelope to open. Callers map this to HTTP 400 so clients can
-	// resubmit the secret after a direct-cutover discard.
-	ErrMissingSecretEnvelope = errors.New("vault credential secret is missing; resubmit the secret")
-)
-
-// credentialBinding builds the AAD binding from a credential's identity. The
-// same binding is used to seal and open, so an envelope cannot be moved to
-// another org/workspace/vault/credential and still decrypt.
-func credentialBinding(credential db.VaultCredential) secrets.Binding {
-	return secrets.Binding{
-		OrganizationUUID:     credential.OrganizationUUID,
-		WorkspaceUUID:        credential.WorkspaceUUID,
-		VaultExternalID:      credential.VaultExternalID,
-		CredentialExternalID: credential.ExternalID,
-	}
-}
-
-// SealCredentialSecret seals a credential's plaintext SecretPayload into
-// SecretEnvelope and drops the plaintext so it is never persisted. Empty or
-// JSON-null payloads are rejected so create/update callers cannot believe a
-// seal succeeded when no envelope was produced. Exported so the platform MCP
-// OAuth callback can seal credentials through the same path as the Vaults API.
-func SealCredentialSecret(ctx context.Context, secretSvc *secrets.Service, credential *db.VaultCredential) error {
-	if len(credential.SecretPayload) == 0 || isJSONNull(credential.SecretPayload) {
-		return errors.New("vault credential secret payload is required to seal")
-	}
-	if err := requireSecretsService(secretSvc); err != nil {
-		return err
-	}
-	envelope, err := secretSvc.Seal(ctx, credentialBinding(*credential), credential.SecretPayload)
-	if err != nil {
-		return fmt.Errorf("seal credential secret: %w", err)
-	}
-	credential.SecretEnvelope = &envelope
-	clearCredentialSecretPayload(credential)
-	return nil
-}
-
-// OpenCredentialSecret decrypts a credential's envelope back into SecretPayload
-// so callers can read the secret for merge or runtime use. A missing envelope
-// returns ErrMissingSecretEnvelope. A present envelope that fails to open
-// fails closed. Decrypted plaintext must not be persisted or logged.
-func OpenCredentialSecret(ctx context.Context, secretSvc *secrets.Service, credential *db.VaultCredential) error {
-	if credential.SecretEnvelope == nil {
-		return ErrMissingSecretEnvelope
-	}
-	if err := requireSecretsService(secretSvc); err != nil {
-		return err
-	}
-	plaintext, err := secretSvc.Open(ctx, credentialBinding(*credential), *credential.SecretEnvelope)
-	if err != nil {
-		return fmt.Errorf("open credential secret: %w", err)
-	}
-	credential.SecretPayload = plaintext
-	return nil
-}
-
-func clearCredentialSecretPayload(credential *db.VaultCredential) {
-	if credential == nil {
-		return
-	}
-	for i := range credential.SecretPayload {
-		credential.SecretPayload[i] = 0
-	}
-	credential.SecretPayload = nil
-}
-
-func requireSecretsService(secretSvc *secrets.Service) error {
-	if secretSvc == nil {
-		return errors.New("vault credential encryption is not configured")
-	}
-	return nil
-}
-
 func writeSecretOpenError(w http.ResponseWriter, r *http.Request, logger *slog.Logger, operation string, err error) {
 	if errors.Is(err, ErrMissingSecretEnvelope) {
 		httpapi.WriteError(w, r, httpapi.NewError(http.StatusBadRequest, "invalid_request_error", ErrMissingSecretEnvelope.Error()))
@@ -256,7 +169,10 @@ func (h *Handler) createVault(w http.ResponseWriter, r *http.Request) {
 		writeBadRequest(w, r, err)
 		return
 	}
-	displayName, err := parseRequiredStringField(fields, "display_name")
+	displayName, err := requiredString(fields, "display_name", "display_name")
+	if err == nil {
+		err = validateDisplayName(displayName)
+	}
 	if err != nil {
 		writeBadRequest(w, r, err)
 		return
@@ -381,11 +297,15 @@ func (h *Handler) updateVaultRoute(w http.ResponseWriter, r *http.Request) {
 	}
 	next := current
 	if raw, ok := fields["display_name"]; ok {
-		next.DisplayName, err = parseRequiredRawString(raw, "display_name")
+		displayName, err := rawString(raw, "display_name")
+		if err == nil {
+			err = validateDisplayName(displayName)
+		}
 		if err != nil {
 			writeBadRequest(w, r, err)
 			return
 		}
+		next.DisplayName = displayName
 	}
 	if raw, ok := fields["metadata"]; ok {
 		next.Metadata, err = patchMetadata(next.Metadata, raw)
@@ -482,7 +402,10 @@ func (h *Handler) createCredentialRoute(w http.ResponseWriter, r *http.Request) 
 		writeBadRequest(w, r, err)
 		return
 	}
-	displayName, err := parseRequiredStringField(fields, "display_name")
+	displayName, err := requiredString(fields, "display_name", "display_name")
+	if err == nil {
+		err = validateDisplayName(displayName)
+	}
 	if err != nil {
 		writeBadRequest(w, r, err)
 		return
@@ -640,11 +563,15 @@ func (h *Handler) updateCredentialRoute(w http.ResponseWriter, r *http.Request) 
 	}
 	next := current
 	if raw, ok := fields["display_name"]; ok {
-		next.DisplayName, err = parseRequiredRawString(raw, "display_name")
+		displayName, err := rawString(raw, "display_name")
+		if err == nil {
+			err = validateDisplayName(displayName)
+		}
 		if err != nil {
 			writeBadRequest(w, r, err)
 			return
 		}
+		next.DisplayName = displayName
 	}
 	if raw, ok := fields["metadata"]; ok {
 		next.Metadata, err = patchMetadata(next.Metadata, raw)
@@ -659,11 +586,14 @@ func (h *Handler) updateCredentialRoute(w http.ResponseWriter, r *http.Request) 
 		// instead of dropping it, then reseal under a fresh DEK.
 		// A missing envelope can still be repaired when the body carries a
 		// complete replacement secret; otherwise ask the client to resubmit.
+		var currentSecret []byte
 		if current.SecretEnvelope != nil {
-			if err := OpenCredentialSecret(r.Context(), h.secretSvc, &current); err != nil {
+			currentSecret, err = openCredentialSecret(r.Context(), h.secretSvc, current)
+			if err != nil {
 				writeSecretOpenError(w, r, h.logger, "update", err)
 				return
 			}
+			defer clear(currentSecret)
 		} else {
 			provides, err := authUpdateProvidesSecretReplacement(current.AuthType, raw)
 			if err != nil {
@@ -675,7 +605,7 @@ func (h *Handler) updateCredentialRoute(w http.ResponseWriter, r *http.Request) 
 				return
 			}
 		}
-		authState, err := normalizeCredentialAuthForUpdate(current, raw)
+		authState, err := normalizeCredentialAuthForUpdate(current, currentSecret, raw)
 		if err != nil {
 			writeBadRequest(w, r, err)
 			return
@@ -800,11 +730,18 @@ func (h *Handler) validateCredentialRoute(w http.ResponseWriter, r *http.Request
 	}
 	// Decrypt transiently to inspect whether a refresh token is present. The
 	// plaintext is not persisted or logged.
-	if err := OpenCredentialSecret(r.Context(), h.secretSvc, &credential); err != nil {
+	plaintext, err := openCredentialSecret(r.Context(), h.secretSvc, credential)
+	if err != nil {
 		writeSecretOpenError(w, r, h.logger, "validate", err)
 		return
 	}
-	hasRefreshToken := hasNestedSecret(credential.SecretPayload, "refresh", "refresh_token")
+	defer clear(plaintext)
+	secret, err := decodeMCPOAuthCredentialSecret(plaintext)
+	if err != nil {
+		writeSecretOpenError(w, r, h.logger, "validate", err)
+		return
+	}
+	hasRefreshToken := secret.Refresh != nil && strings.TrimSpace(secret.Refresh.RefreshToken) != ""
 	refreshStatus := "no_refresh_token"
 	if hasRefreshToken {
 		refreshStatus = "connect_error"
@@ -923,31 +860,6 @@ func decodeObjectBody(w http.ResponseWriter, r *http.Request) (map[string]json.R
 	return fields, nil
 }
 
-func parseRequiredStringField(fields map[string]json.RawMessage, name string) (string, error) {
-	raw, ok := fields[name]
-	if !ok {
-		return "", fmt.Errorf("%s is required", name)
-	}
-	return parseRequiredRawString(raw, name)
-}
-
-func parseRequiredRawString(raw json.RawMessage, name string) (string, error) {
-	if isJSONNull(raw) {
-		return "", fmt.Errorf("%s cannot be null", name)
-	}
-	var value string
-	if err := json.Unmarshal(raw, &value); err != nil {
-		return "", fmt.Errorf("%s must be a string", name)
-	}
-	if strings.TrimSpace(value) == "" {
-		return "", fmt.Errorf("%s must be non-empty", name)
-	}
-	if len(value) > 255 {
-		return "", fmt.Errorf("%s must be at most 255 characters", name)
-	}
-	return value, nil
-}
-
 func normalizeMetadata(raw json.RawMessage) (json.RawMessage, error) {
 	if isJSONNull(raw) {
 		return json.RawMessage(`{}`), nil
@@ -1004,594 +916,11 @@ func validateMetadata(metadata map[string]string) error {
 	return nil
 }
 
-func normalizeCredentialAuthForCreate(raw json.RawMessage) (credentialAuthState, error) {
-	fields, err := objectFromRaw(raw, "auth")
-	if err != nil {
-		return credentialAuthState{}, err
-	}
-	authType, err := requiredString(fields, "type", "auth.type")
-	if err != nil {
-		return credentialAuthState{}, err
-	}
-	switch authType {
-	case "mcp_oauth":
-		return normalizeMCPOAuthForCreate(fields)
-	case "static_bearer":
-		return normalizeStaticBearerForCreate(fields)
-	case "environment_variable":
-		return normalizeEnvironmentVariableForCreate(fields)
-	default:
-		return credentialAuthState{}, errors.New("auth.type must be mcp_oauth, static_bearer, or environment_variable")
-	}
-}
-
-func normalizeMCPOAuthForCreate(fields map[string]json.RawMessage) (credentialAuthState, error) {
-	serverURL, err := requiredString(fields, "mcp_server_url", "auth.mcp_server_url")
-	if err != nil {
-		return credentialAuthState{}, err
-	}
-	if err := validateHTTPURL(serverURL, "auth.mcp_server_url"); err != nil {
-		return credentialAuthState{}, err
-	}
-	accessToken, err := requiredString(fields, "access_token", "auth.access_token")
-	if err != nil {
-		return credentialAuthState{}, err
-	}
-	publicAuth := map[string]any{
-		"type":           "mcp_oauth",
-		"mcp_server_url": serverURL,
-	}
-	secretPayload := map[string]any{
-		"type":         "mcp_oauth",
-		"access_token": accessToken,
-	}
-	if expiresAt, ok, err := optionalString(fields, "expires_at", "auth.expires_at"); err != nil {
-		return credentialAuthState{}, err
-	} else if ok {
-		if err := validateRFC3339(expiresAt, "auth.expires_at"); err != nil {
-			return credentialAuthState{}, err
-		}
-		publicAuth["expires_at"] = expiresAt
-	}
-	if rawRefresh, ok := fields["refresh"]; ok && !isJSONNull(rawRefresh) {
-		publicRefresh, secretRefresh, err := normalizeMCPOAuthRefreshForCreate(rawRefresh)
-		if err != nil {
-			return credentialAuthState{}, err
-		}
-		publicAuth["refresh"] = publicRefresh
-		secretPayload["refresh"] = secretRefresh
-	}
-	return credentialAuthStateFromMaps("mcp_oauth", serverURL, publicAuth, secretPayload)
-}
-
-func normalizeMCPOAuthRefreshForCreate(raw json.RawMessage) (map[string]any, map[string]any, error) {
-	fields, err := objectFromRaw(raw, "auth.refresh")
-	if err != nil {
-		return nil, nil, err
-	}
-	tokenEndpoint, err := requiredString(fields, "token_endpoint", "auth.refresh.token_endpoint")
-	if err != nil {
-		return nil, nil, err
-	}
-	if err := validateHTTPURL(tokenEndpoint, "auth.refresh.token_endpoint"); err != nil {
-		return nil, nil, err
-	}
-	clientID, err := requiredString(fields, "client_id", "auth.refresh.client_id")
-	if err != nil {
-		return nil, nil, err
-	}
-	refreshToken, err := requiredString(fields, "refresh_token", "auth.refresh.refresh_token")
-	if err != nil {
-		return nil, nil, err
-	}
-	publicTokenAuth, secretTokenAuth, err := normalizeTokenEndpointAuthForCreate(fields["token_endpoint_auth"])
-	if err != nil {
-		return nil, nil, err
-	}
-	publicRefresh := map[string]any{
-		"token_endpoint":      tokenEndpoint,
-		"client_id":           clientID,
-		"token_endpoint_auth": publicTokenAuth,
-	}
-	secretRefresh := map[string]any{
-		"refresh_token":       refreshToken,
-		"token_endpoint_auth": secretTokenAuth,
-	}
-	if value, ok, err := optionalString(fields, "scope", "auth.refresh.scope"); err != nil {
-		return nil, nil, err
-	} else if ok {
-		publicRefresh["scope"] = value
-	}
-	if value, ok, err := optionalString(fields, "resource", "auth.refresh.resource"); err != nil {
-		return nil, nil, err
-	} else if ok {
-		publicRefresh["resource"] = value
-	}
-	return publicRefresh, secretRefresh, nil
-}
-
-func normalizeTokenEndpointAuthForCreate(raw json.RawMessage) (map[string]any, map[string]any, error) {
-	if len(raw) == 0 || isJSONNull(raw) {
-		return map[string]any{"type": "none"}, map[string]any{"type": "none"}, nil
-	}
-	fields, err := objectFromRaw(raw, "auth.refresh.token_endpoint_auth")
-	if err != nil {
-		return nil, nil, err
-	}
-	authType, err := requiredString(fields, "type", "auth.refresh.token_endpoint_auth.type")
-	if err != nil {
-		return nil, nil, err
-	}
-	switch authType {
-	case "none":
-		return map[string]any{"type": "none"}, map[string]any{"type": "none"}, nil
-	case "client_secret_basic", "client_secret_post":
-		clientSecret, err := requiredString(fields, "client_secret", "auth.refresh.token_endpoint_auth.client_secret")
-		if err != nil {
-			return nil, nil, err
-		}
-		return map[string]any{"type": authType}, map[string]any{"type": authType, "client_secret": clientSecret}, nil
-	default:
-		return nil, nil, errors.New("auth.refresh.token_endpoint_auth.type must be none, client_secret_basic, or client_secret_post")
-	}
-}
-
-func normalizeStaticBearerForCreate(fields map[string]json.RawMessage) (credentialAuthState, error) {
-	serverURL, err := requiredString(fields, "mcp_server_url", "auth.mcp_server_url")
-	if err != nil {
-		return credentialAuthState{}, err
-	}
-	if err := validateHTTPURL(serverURL, "auth.mcp_server_url"); err != nil {
-		return credentialAuthState{}, err
-	}
-	token, err := requiredString(fields, "token", "auth.token")
-	if err != nil {
-		return credentialAuthState{}, err
-	}
-	publicAuth := map[string]any{"type": "static_bearer", "mcp_server_url": serverURL}
-	secretPayload := map[string]any{"type": "static_bearer", "token": token}
-	return credentialAuthStateFromMaps("static_bearer", serverURL, publicAuth, secretPayload)
-}
-
-func normalizeEnvironmentVariableForCreate(fields map[string]json.RawMessage) (credentialAuthState, error) {
-	secretName, err := requiredString(fields, "secret_name", "auth.secret_name")
-	if err != nil {
-		return credentialAuthState{}, err
-	}
-	if err := validateSecretName(secretName); err != nil {
-		return credentialAuthState{}, err
-	}
-	secretValue, err := requiredString(fields, "secret_value", "auth.secret_value")
-	if err != nil {
-		return credentialAuthState{}, err
-	}
-	networking, err := normalizeCredentialNetworking(fields["networking"])
-	if err != nil {
-		return credentialAuthState{}, err
-	}
-	publicAuth := map[string]any{
-		"type":        "environment_variable",
-		"secret_name": secretName,
-		"networking":  networking,
-	}
-	secretPayload := map[string]any{"type": "environment_variable", "secret_value": secretValue}
-	return credentialAuthStateFromMaps("environment_variable", secretName, publicAuth, secretPayload)
-}
-
-func normalizeCredentialAuthForUpdate(current db.VaultCredential, raw json.RawMessage) (credentialAuthState, error) {
-	fields, err := objectFromRaw(raw, "auth")
-	if err != nil {
-		return credentialAuthState{}, err
-	}
-	authType, err := requiredString(fields, "type", "auth.type")
-	if err != nil {
-		return credentialAuthState{}, err
-	}
-	if authType != current.AuthType {
-		return credentialAuthState{}, errors.New("auth.type cannot be changed")
-	}
-	publicAuth := rawObjectMap(current.Auth)
-	secretPayload := rawObjectMap(current.SecretPayload)
-	if secretPayload == nil {
-		secretPayload = map[string]any{"type": current.AuthType}
-	}
-	publicAuth["type"] = current.AuthType
-	secretPayload["type"] = current.AuthType
-
-	switch current.AuthType {
-	case "mcp_oauth":
-		if _, ok := fields["mcp_server_url"]; ok {
-			return credentialAuthState{}, errors.New("auth.mcp_server_url is immutable")
-		}
-		if rawAccessToken, ok := fields["access_token"]; ok {
-			accessToken, err := rawString(rawAccessToken, "auth.access_token")
-			if err != nil {
-				return credentialAuthState{}, err
-			}
-			secretPayload["access_token"] = accessToken
-		}
-		if rawExpiresAt, ok := fields["expires_at"]; ok {
-			expiresAt, err := rawString(rawExpiresAt, "auth.expires_at")
-			if err != nil {
-				return credentialAuthState{}, err
-			}
-			if err := validateRFC3339(expiresAt, "auth.expires_at"); err != nil {
-				return credentialAuthState{}, err
-			}
-			publicAuth["expires_at"] = expiresAt
-		}
-		if rawRefresh, ok := fields["refresh"]; ok {
-			if isJSONNull(rawRefresh) {
-				delete(publicAuth, "refresh")
-				delete(secretPayload, "refresh")
-			} else if err := patchMCPOAuthRefreshForUpdate(publicAuth, secretPayload, rawRefresh); err != nil {
-				return credentialAuthState{}, err
-			}
-		}
-		return credentialAuthStateFromMaps(current.AuthType, current.CredentialKey, publicAuth, secretPayload)
-	case "static_bearer":
-		key := current.CredentialKey
-		if rawServerURL, ok := fields["mcp_server_url"]; ok {
-			serverURL, err := rawString(rawServerURL, "auth.mcp_server_url")
-			if err != nil {
-				return credentialAuthState{}, err
-			}
-			if err := validateHTTPURL(serverURL, "auth.mcp_server_url"); err != nil {
-				return credentialAuthState{}, err
-			}
-			publicAuth["mcp_server_url"] = serverURL
-			key = serverURL
-		}
-		if rawToken, ok := fields["token"]; ok {
-			token, err := rawString(rawToken, "auth.token")
-			if err != nil {
-				return credentialAuthState{}, err
-			}
-			secretPayload["token"] = token
-		}
-		return credentialAuthStateFromMaps(current.AuthType, key, publicAuth, secretPayload)
-	case "environment_variable":
-		if _, ok := fields["secret_name"]; ok {
-			return credentialAuthState{}, errors.New("auth.secret_name is immutable")
-		}
-		if rawSecretValue, ok := fields["secret_value"]; ok {
-			secretValue, err := rawString(rawSecretValue, "auth.secret_value")
-			if err != nil {
-				return credentialAuthState{}, err
-			}
-			secretPayload["secret_value"] = secretValue
-		}
-		if rawNetworking, ok := fields["networking"]; ok {
-			networking, err := normalizeCredentialNetworking(rawNetworking)
-			if err != nil {
-				return credentialAuthState{}, err
-			}
-			publicAuth["networking"] = networking
-		}
-		return credentialAuthStateFromMaps(current.AuthType, current.CredentialKey, publicAuth, secretPayload)
-	default:
-		return credentialAuthState{}, errors.New("stored credential auth type is invalid")
-	}
-}
-
-// authUpdateProvidesSecretReplacement reports whether an auth update body
-// carries enough secret material to reseal without opening an existing
-// envelope (used to repair credentials that lost their envelope).
-func authUpdateProvidesSecretReplacement(authType string, raw json.RawMessage) (bool, error) {
-	fields, err := objectFromRaw(raw, "auth")
-	if err != nil {
-		return false, err
-	}
-	switch authType {
-	case "static_bearer":
-		_, ok := fields["token"]
-		return ok, nil
-	case "environment_variable":
-		_, ok := fields["secret_value"]
-		return ok, nil
-	case "mcp_oauth":
-		_, ok := fields["access_token"]
-		return ok, nil
-	default:
-		return false, nil
-	}
-}
-
-func patchMCPOAuthRefreshForUpdate(publicAuth, secretPayload map[string]any, raw json.RawMessage) error {
-	fields, err := objectFromRaw(raw, "auth.refresh")
-	if err != nil {
-		return err
-	}
-	if _, ok := fields["token_endpoint"]; ok {
-		return errors.New("auth.refresh.token_endpoint is immutable")
-	}
-	if _, ok := fields["client_id"]; ok {
-		return errors.New("auth.refresh.client_id is immutable")
-	}
-	if _, ok := fields["resource"]; ok {
-		return errors.New("auth.refresh.resource is immutable")
-	}
-	publicRefresh := nestedMap(publicAuth, "refresh")
-	secretRefresh := nestedMap(secretPayload, "refresh")
-	if publicRefresh == nil {
-		return errors.New("auth.refresh cannot be added after creation")
-	}
-	if secretRefresh == nil {
-		secretRefresh = map[string]any{}
-	}
-	if rawRefreshToken, ok := fields["refresh_token"]; ok {
-		refreshToken, err := rawString(rawRefreshToken, "auth.refresh.refresh_token")
-		if err != nil {
-			return err
-		}
-		secretRefresh["refresh_token"] = refreshToken
-	}
-	if rawScope, ok := fields["scope"]; ok {
-		if isJSONNull(rawScope) {
-			delete(publicRefresh, "scope")
-		} else {
-			scope, err := rawString(rawScope, "auth.refresh.scope")
-			if err != nil {
-				return err
-			}
-			publicRefresh["scope"] = scope
-		}
-	}
-	if rawTokenAuth, ok := fields["token_endpoint_auth"]; ok {
-		publicTokenAuth, secretTokenAuth, err := normalizeTokenEndpointAuthForUpdate(rawTokenAuth)
-		if err != nil {
-			return err
-		}
-		publicRefresh["token_endpoint_auth"] = publicTokenAuth
-		secretRefresh["token_endpoint_auth"] = secretTokenAuth
-	}
-	publicAuth["refresh"] = publicRefresh
-	secretPayload["refresh"] = secretRefresh
-	return nil
-}
-
-func normalizeTokenEndpointAuthForUpdate(raw json.RawMessage) (map[string]any, map[string]any, error) {
-	if isJSONNull(raw) {
-		return map[string]any{"type": "none"}, map[string]any{"type": "none"}, nil
-	}
-	fields, err := objectFromRaw(raw, "auth.refresh.token_endpoint_auth")
-	if err != nil {
-		return nil, nil, err
-	}
-	authType, err := requiredString(fields, "type", "auth.refresh.token_endpoint_auth.type")
-	if err != nil {
-		return nil, nil, err
-	}
-	switch authType {
-	case "none":
-		return map[string]any{"type": "none"}, map[string]any{"type": "none"}, nil
-	case "client_secret_basic", "client_secret_post":
-		clientSecret, err := requiredString(fields, "client_secret", "auth.refresh.token_endpoint_auth.client_secret")
-		if err != nil {
-			return nil, nil, err
-		}
-		return map[string]any{"type": authType}, map[string]any{"type": authType, "client_secret": clientSecret}, nil
-	default:
-		return nil, nil, errors.New("auth.refresh.token_endpoint_auth.type must be none, client_secret_basic, or client_secret_post")
-	}
-}
-
-func normalizeCredentialNetworking(raw json.RawMessage) (map[string]any, error) {
-	if len(raw) == 0 || isJSONNull(raw) {
-		return map[string]any{"type": "unrestricted"}, nil
-	}
-	fields, err := objectFromRaw(raw, "auth.networking")
-	if err != nil {
-		return nil, err
-	}
-	networkType := rawStringOrEmpty(fields["type"])
-	if networkType == "" {
-		networkType = "unrestricted"
-	}
-	switch networkType {
-	case "unrestricted":
-		return map[string]any{"type": "unrestricted"}, nil
-	case "limited":
-		hosts := []string{}
-		if rawHosts, ok := fields["allowed_hosts"]; ok && !isJSONNull(rawHosts) {
-			values, err := stringArray(rawHosts, "auth.networking.allowed_hosts")
-			if err != nil {
-				return nil, err
-			}
-			if len(values) > 16 {
-				return nil, errors.New("auth.networking.allowed_hosts must contain at most 16 hosts")
-			}
-			for _, host := range values {
-				if err := validateCredentialHost(host); err != nil {
-					return nil, err
-				}
-			}
-			hosts = values
-		}
-		return map[string]any{"type": "limited", "allowed_hosts": hosts}, nil
-	default:
-		return nil, errors.New("auth.networking.type must be unrestricted or limited")
-	}
-}
-
-func credentialAuthStateFromMaps(authType, key string, publicAuth, secretPayload map[string]any) (credentialAuthState, error) {
-	publicRaw, err := marshalRaw(publicAuth)
-	if err != nil {
-		return credentialAuthState{}, err
-	}
-	secretRaw, err := marshalRaw(secretPayload)
-	if err != nil {
-		return credentialAuthState{}, err
-	}
-	return credentialAuthState{
-		AuthType:      authType,
-		Key:           key,
-		PublicAuth:    publicRaw,
-		SecretPayload: secretRaw,
-	}, nil
-}
-
-func objectFromRaw(raw json.RawMessage, name string) (map[string]json.RawMessage, error) {
-	if len(raw) == 0 {
-		return nil, fmt.Errorf("%s is required", name)
-	}
-	if isJSONNull(raw) {
-		return nil, fmt.Errorf("%s cannot be null", name)
-	}
-	var fields map[string]json.RawMessage
-	if err := json.Unmarshal(raw, &fields); err != nil {
-		return nil, fmt.Errorf("%s must be an object", name)
-	}
-	if fields == nil {
-		return nil, fmt.Errorf("%s must be an object", name)
-	}
-	return fields, nil
-}
-
-func requiredString(fields map[string]json.RawMessage, key, name string) (string, error) {
-	raw, ok := fields[key]
-	if !ok {
-		return "", fmt.Errorf("%s is required", name)
-	}
-	return rawString(raw, name)
-}
-
-func optionalString(fields map[string]json.RawMessage, key, name string) (string, bool, error) {
-	raw, ok := fields[key]
-	if !ok || isJSONNull(raw) {
-		return "", false, nil
-	}
-	value, err := rawString(raw, name)
-	if err != nil {
-		return "", false, err
-	}
-	return value, true, nil
-}
-
-func rawString(raw json.RawMessage, name string) (string, error) {
-	if isJSONNull(raw) {
-		return "", fmt.Errorf("%s cannot be null", name)
-	}
-	var value string
-	if err := json.Unmarshal(raw, &value); err != nil {
-		return "", fmt.Errorf("%s must be a string", name)
-	}
-	if strings.TrimSpace(value) == "" {
-		return "", fmt.Errorf("%s must be non-empty", name)
-	}
-	return value, nil
-}
-
-func rawStringOrEmpty(raw json.RawMessage) string {
-	if len(raw) == 0 || isJSONNull(raw) {
-		return ""
-	}
-	var value string
-	if err := json.Unmarshal(raw, &value); err != nil {
-		return ""
-	}
-	return value
-}
-
-func stringArray(raw json.RawMessage, name string) ([]string, error) {
-	var values []string
-	if err := json.Unmarshal(raw, &values); err != nil {
-		return nil, fmt.Errorf("%s must be an array of strings", name)
-	}
-	for _, value := range values {
-		if strings.TrimSpace(value) == "" {
-			return nil, fmt.Errorf("%s entries must be non-empty strings", name)
-		}
-		if len(value) > 253 {
-			return nil, fmt.Errorf("%s entries must be at most 253 characters", name)
-		}
-	}
-	return values, nil
-}
-
-func validateHTTPURL(value, name string) error {
-	parsed, err := url.Parse(value)
-	if err != nil || parsed.Scheme == "" || parsed.Host == "" {
-		return fmt.Errorf("%s must be a valid URL", name)
-	}
-	if parsed.Scheme != "https" && parsed.Scheme != "http" {
-		return fmt.Errorf("%s must use http or https", name)
+func validateDisplayName(displayName string) error {
+	if len(displayName) > 255 {
+		return errors.New("display_name must be at most 255 characters")
 	}
 	return nil
-}
-
-func validateRFC3339(value, name string) error {
-	if _, err := time.Parse(time.RFC3339Nano, value); err != nil {
-		return fmt.Errorf("%s must be RFC3339", name)
-	}
-	return nil
-}
-
-func validateSecretName(value string) error {
-	if len(value) > 255 {
-		return errors.New("auth.secret_name must be at most 255 characters")
-	}
-	return nil
-}
-
-func validateCredentialHost(host string) error {
-	if strings.Contains(host, "://") || strings.Contains(host, "/") || strings.Contains(host, ":") || strings.Contains(host, "[") || strings.Contains(host, "]") {
-		return errors.New("auth.networking.allowed_hosts entries must be hostnames without URL schemes")
-	}
-	if len(host) > 253 {
-		return errors.New("auth.networking.allowed_hosts entries must be at most 253 characters")
-	}
-	if !credentialHostPattern.MatchString(host) {
-		return errors.New("auth.networking.allowed_hosts entries must be valid hostnames")
-	}
-	return nil
-}
-
-func rawObjectMap(raw json.RawMessage) map[string]any {
-	var value map[string]any
-	if len(raw) == 0 || isJSONNull(raw) {
-		return map[string]any{}
-	}
-	if err := json.Unmarshal(raw, &value); err != nil || value == nil {
-		return map[string]any{}
-	}
-	return value
-}
-
-func nestedMap(parent map[string]any, key string) map[string]any {
-	value, ok := parent[key]
-	if !ok || value == nil {
-		return nil
-	}
-	if mapped, ok := value.(map[string]any); ok {
-		return mapped
-	}
-	raw, err := json.Marshal(value)
-	if err != nil {
-		return nil
-	}
-	var mapped map[string]any
-	if err := json.Unmarshal(raw, &mapped); err != nil {
-		return nil
-	}
-	return mapped
-}
-
-func hasNestedSecret(raw json.RawMessage, parent, child string) bool {
-	root := rawObjectMap(raw)
-	nested := nestedMap(root, parent)
-	if nested == nil {
-		return false
-	}
-	value, ok := nested[child].(string)
-	return ok && strings.TrimSpace(value) != ""
-}
-
-func fieldOrDefault(fields map[string]json.RawMessage, name, fallback string) json.RawMessage {
-	if raw, ok := fields[name]; ok {
-		return raw
-	}
-	return json.RawMessage(fallback)
 }
 
 func parseLimit(r *http.Request) (int, error) {
@@ -1673,18 +1002,6 @@ func decodeCredentialCursor(raw string) (*db.VaultCredentialPageCursor, error) {
 		return nil, errors.New("page is invalid")
 	}
 	return &db.VaultCredentialPageCursor{CreatedAt: createdAt.UTC(), UUID: payload.UUID}, nil
-}
-
-func isJSONNull(raw json.RawMessage) bool {
-	return strings.TrimSpace(string(raw)) == "null"
-}
-
-func marshalRaw(value any) (json.RawMessage, error) {
-	data, err := json.Marshal(value)
-	if err != nil {
-		return nil, err
-	}
-	return json.RawMessage(data), nil
 }
 
 func rawOr(raw json.RawMessage, fallback string) json.RawMessage {
