+++ internal/vaults/injector_test.go
@@ -3,7 +3,6 @@ package vaults
 import (
 	"context"
 	"encoding/json"
-	"net/http"
 	"testing"
 
 	"github.com/superduck-ai/open-managed-agents/internal/db"
@@ -12,6 +11,7 @@ import (
 
 func TestOpenStaticBearerToken(t *testing.T) {
 	svc := newTestSecretsService(t)
+	injector := &Injector{secretSvc: svc}
 	credential := &db.VaultCredential{
 		OrganizationUUID: "00000000-0000-0000-0000-000000000001",
 		WorkspaceUUID:    "00000000-0000-0000-0000-000000000002",
@@ -26,48 +26,25 @@ func TestOpenStaticBearerToken(t *testing.T) {
 	if err := SealCredentialSecret(context.Background(), svc, credential); err != nil {
 		t.Fatalf("seal: %v", err)
 	}
+	credential.SecretPayload = json.RawMessage(`{"caller":"owned"}`)
 
-	token, err := openStaticBearerToken(context.Background(), svc, credential)
+	token, err := injector.openStaticBearerToken(context.Background(), credential)
 	if err != nil {
 		t.Fatalf("openStaticBearerToken: %v", err)
 	}
 	if token != "super-secret-token" {
 		t.Fatalf("token = %q", token)
 	}
-	if len(credential.SecretPayload) != 0 {
-		t.Fatal("plaintext should be cleared after openStaticBearerToken")
+	if string(credential.SecretPayload) != `{"caller":"owned"}` {
+		t.Fatalf("caller payload was modified: %s", credential.SecretPayload)
 	}
 }
 
-func TestOpenStaticBearerTokenErrors(t *testing.T) {
+func TestOpenStaticBearerTokenMissingEnvelope(t *testing.T) {
 	svc := newTestSecretsService(t)
-	if _, err := openStaticBearerToken(context.Background(), svc, &db.VaultCredential{ExternalID: "cred_missing"}); err == nil {
+	injector := &Injector{secretSvc: svc}
+	if _, err := injector.openStaticBearerToken(context.Background(), &db.VaultCredential{ExternalID: "cred_missing"}); err == nil {
 		t.Fatal("expected error for missing envelope")
-	}
-	if _, err := openStaticBearerToken(context.Background(), nil, &db.VaultCredential{
-		ExternalID:     "cred_x",
-		SecretEnvelope: &secrets.Envelope{Ciphertext: []byte("x"), Nonce: []byte("n"), WrappedDEK: []byte("w"), FormatVersion: 1, KeyProvider: "local", KeyVersion: 1},
-	}); err == nil {
-		t.Fatal("expected error when secrets service is nil")
-	}
-}
-
-func TestRewriteAuthorizationNilInjector(t *testing.T) {
-	var injector *Injector
-	header := make(http.Header)
-	header.Set("Authorization", "Bearer client-token")
-	if err := injector.RewriteAuthorization(
-		context.Background(),
-		"cse_test",
-		"00000000-0000-0000-0000-000000000001",
-		"00000000-0000-0000-0000-000000000002",
-		mustURL(t, "https://mcp.example.com/mcp"),
-		header,
-	); err != nil {
-		t.Fatalf("RewriteAuthorization: %v", err)
-	}
-	if got := header.Get("Authorization"); got != "Bearer client-token" {
-		t.Fatalf("Authorization = %q, want client token preserved", got)
 	}
 }
